requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.7.4.post0
    # via
    #   discord-py
    #   gcsfs
async-generator==1.10
    # via
    #   trio
    #   trio-websocket
async-timeout==3.0.1
    # via aiohttp
attrs==21.4.0
    # via
    #   aiohttp
    #   outcome
    #   trio
beautifulsoup4==4.10.0
    # via -r requirements.in
blinker==1.4
    # via selenium-wire
boto3==1.20.37
    # via -r requirements.in
botocore==1.23.37
    # via
    #   boto3
    #   s3transfer
brotli==1.0.9
    # via selenium-wire
cachetools==4.2.4
    # via google-auth
certifi==2021.10.8
    # via
    #   requests
    #   selenium-wire
    #   urllib3
cffi==1.15.0
    # via cryptography
chardet==4.0.0
    # via
    #   aiohttp
    #   requests
click==8.0.3
    # via
    #   distributed
    #   prefect
cloudpickle==2.0.0
    # via
    #   dask
    #   distributed
    #   prefect
colorama==0.4.4
    # via crayons
configparser==5.2.0
    # via webdriver-manager
cramjam==2.5.0
    # via fastparquet
crayons==0.4.0
    # via webdriver-manager
croniter==1.2.0
    # via prefect
cryptography==36.0.1
    # via
    #   pyopenssl
    #   urllib3
dask==2022.1.0
    # via
    #   distributed
    #   prefect
decorator==5.1.1
    # via gcsfs
discord-py==1.7.3
    # via -r requirements.in
distributed==2022.1.0
    # via prefect
docker==5.0.3
    # via prefect
fastparquet==0.7.2
    # via -r requirements.in
fsspec==2022.5.0
    # via
    #   dask
    #   fastparquet
    #   gcsfs
gcsfs==2022.5.0
    # via -r requirements.in
google-api-core[grpc]==2.4.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.3.3
    # via
    #   -r requirements.in
    #   gcsfs
    #   google-api-core
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
    #   pandas-gbq
    #   pydata-google-auth
google-auth-oauthlib==0.4.6
    # via
    #   gcsfs
    #   pandas-gbq
    #   pydata-google-auth
google-cloud-bigquery[bqstorage,pandas]==2.32.0
    # via pandas-gbq
google-cloud-bigquery-storage==2.11.0
    # via google-cloud-bigquery
google-cloud-core==2.3.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-storage==2.3.0
    # via gcsfs
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.3.3
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos==1.54.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.43.0
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   grpcio-status
grpcio-status==1.43.0
    # via google-api-core
h11==0.12.0
    # via wsproto
h2==4.1.0
    # via selenium-wire
heapdict==1.0.1
    # via zict
hpack==4.0.0
    # via h2
hyperframe==6.0.1
    # via
    #   h2
    #   selenium-wire
idna==2.10
    # via
    #   requests
    #   trio
    #   urllib3
    #   yarl
jinja2==3.0.3
    # via distributed
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
kaitaistruct==0.9
    # via selenium-wire
libcst==0.4.0
    # via google-cloud-bigquery-storage
locket==0.2.1
    # via partd
markupsafe==2.0.1
    # via jinja2
marshmallow==3.14.1
    # via
    #   marshmallow-oneofschema
    #   prefect
marshmallow-oneofschema==3.0.1
    # via prefect
msgpack==1.0.3
    # via
    #   distributed
    #   prefect
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via
    #   prefect
    #   typing-inspect
numpy==1.22.1
    # via
    #   fastparquet
    #   pandas
    #   pandas-gbq
    #   pyarrow
oauthlib==3.1.1
    # via requests-oauthlib
outcome==1.1.0
    # via trio
packaging==21.3
    # via
    #   dask
    #   distributed
    #   google-cloud-bigquery
    #   prefect
pandas==1.3.5
    # via
    #   -r requirements.in
    #   fastparquet
    #   google-cloud-bigquery
    #   pandas-gbq
pandas-gbq==0.16.0
    # via -r requirements.in
partd==1.2.0
    # via dask
pendulum==2.1.2
    # via prefect
prefect==0.15.12
    # via -r requirements.in
proto-plus==1.19.8
    # via
    #   google-cloud-bigquery
    #   google-cloud-bigquery-storage
protobuf==3.19.3
    # via
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psutil==5.9.0
    # via distributed
pyarrow==6.0.1
    # via
    #   -r requirements.in
    #   google-cloud-bigquery
    #   pandas-gbq
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
    #   selenium-wire
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pydantic==1.9.0
    # via -r requirements.in
pydata-google-auth==1.3.0
    # via pandas-gbq
pymongo==4.1.1
    # via -r requirements.in
pyopenssl==21.0.0
    # via
    #   selenium-wire
    #   urllib3
pyparsing==3.0.6
    # via
    #   packaging
    #   selenium-wire
pysocks==1.7.1
    # via selenium-wire
python-box==5.4.1
    # via prefect
python-dateutil==2.8.2
    # via
    #   botocore
    #   croniter
    #   google-cloud-bigquery
    #   pandas
    #   pendulum
    #   prefect
python-dotenv==0.19.2
    # via -r requirements.in
python-slugify==5.0.2
    # via prefect
pytz==2021.3
    # via
    #   pandas
    #   prefect
pytzdata==2020.1
    # via pendulum
pyyaml==6.0
    # via
    #   dask
    #   distributed
    #   libcst
    #   prefect
requests==2.25.1
    # via
    #   -r requirements.in
    #   docker
    #   gcsfs
    #   google-api-core
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   prefect
    #   requests-oauthlib
    #   tweepy
    #   undetected-chromedriver
    #   webdriver-manager
requests-oauthlib==1.3.0
    # via
    #   google-auth-oauthlib
    #   tweepy
rsa==4.8
    # via google-auth
s3transfer==0.5.0
    # via boto3
selenium==4.1.0
    # via
    #   -r requirements.in
    #   selenium-wire
    #   undetected-chromedriver
selenium-wire==4.6.0
    # via -r requirements.in
six==1.16.0
    # via
    #   google-auth
    #   grpcio
    #   pyopenssl
    #   python-dateutil
    #   thrift
sniffio==1.2.0
    # via trio
sortedcontainers==2.4.0
    # via
    #   distributed
    #   trio
soupsieve==2.3.1
    # via beautifulsoup4
tabulate==0.8.9
    # via prefect
tblib==1.7.0
    # via distributed
text-unidecode==1.3
    # via python-slugify
thrift==0.15.0
    # via fastparquet
toml==0.10.2
    # via prefect
toolz==0.11.2
    # via
    #   dask
    #   distributed
    #   partd
tornado==6.1
    # via distributed
trio==0.19.0
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.9.2
    # via selenium
tweepy==4.4.0
    # via -r requirements.in
types-requests==2.27.7
    # via -r requirements.in
types-urllib3==1.26.7
    # via types-requests
typing-extensions==4.0.1
    # via
    #   aiohttp
    #   libcst
    #   pydantic
    #   typing-inspect
typing-inspect==0.7.1
    # via libcst
undetected-chromedriver==3.1.3
    # via -r requirements.in
urllib3[secure]==1.26.8
    # via
    #   botocore
    #   prefect
    #   requests
    #   selenium
webdriver-manager==3.5.2
    # via -r requirements.in
websocket-client==1.2.3
    # via docker
websockets==9.1
    # via undetected-chromedriver
wsproto==1.0.0
    # via
    #   selenium-wire
    #   trio-websocket
yarl==1.7.2
    # via aiohttp
zict==2.0.0
    # via distributed
zstandard==0.17.0
    # via selenium-wire

# The following packages are considered to be unsafe in a requirements file:
# setuptools
